requirements.txt
Django==1.6.5
Pillow==2.5.1
South==1.0
django-bootstrap-form==3.1
django-registration==1.0
pip==1.5.6
requests==2.3.0
setuptools==5.4.1
